1requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
minimalmodbus==2.1.1
modbus_tk==1.1.5
numpy==2.2.6
paho-mqtt==2.1.0
pillow==11.2.1
psutil==7.0.0
pyserial==3.5
requests==2.32.3
RPi.GPIO==0.7.1
smbus==1.1.post2
smbus2==0.5.0
spidev==3.7
urllib3==2.4.0
